requirements.txt
# YOLOv11 및 딥러닝 프레임워크
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# 데이터 처리
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 시각화 및 분석
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Jupyter Lab
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# 유틸리티
tqdm>=4.65.0
pycocotools>=2.0.6

# 데이터 증강 (선택사항)
albumentations>=1.3.0

